requirements.txt
# 🎯 المتطلبات الأساسية لنظام HPO

# --- المحرك الرئيسي للتحسين ---
# Optuna: لإدارة عمليات البحث، السامبلرات (TPE, GP)، والـ pruners (ASHA)
optuna

# --- أطر عمل تعلم الآلة المدعومة ---
# Scikit-learn: لتحسين نماذج تعلم الآلة التقليدية
scikit-learn

# XGBoost: لتحسين نماذج الـ Gradient Boosting
xgboost

# PyTorch: لتحسين نماذج التعلم العميق
# ملاحظة: قد تحتاج إلى تثبيت نسخة معينة (CPU/GPU) يدويًا من موقع PyTorch الرسمي
# torch
# torchvision
# torchaudio

# --- مكتبات مساعدة ---
# NumPy: للعمليات الرياضية والمصفوفات
numpy

# Pandas: لمعالجة البيانات والجداول
pandas

# Plotly: لإنشاء الرسوم البيانية التفاعلية (يستخدمها Optuna)
plotly

# PyYAML: لقراءة ملفات الإعدادات (configs)
PyYAML

# --- أدوات جودة الكود (اختياري لكن موصى به) ---
# flake8
# black
# mypy